refactor down to one file, as the implementation is the same
